fix: 대쉬보드 페이지 카드 링크들의 링크가 정상적으로 동작하지 않던 점 수정
